// src/api/generated/model/profileResponse.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * 캡스톤 백엔드
 * 스웨거 API 명세.
 * OpenAPI spec version: v1.0.0
 */

export interface ProfileResponse {
  /** 사용자 이름 */
  name?: string;
  /** 사용자 이메일 */
  email?: string;
  /** 전화번호 (미설정 시 null) */
  phone?: string;
  /** GitHub 프로필 URL (미설정 시 null) */
  github?: string;
  /** 기술 블로그 URL (미설정 시 null) */
  blog?: string;
  /** 프로필 이미지 URL (미설정 시 null) */
  profileImage?: string;
}